[FE/BE]feat: 마이페이지 이벤트 처리

+++ backend/src/main/java/com/example/backend/dto/UserMyPageResponseDto.java
@@ -16,7 +16,7 @@ public class UserMyPageResponseDto {
     private String location;
     private LocalDateTime createdAt;
 
-    private int groups;
-    private int events;
-    private int qna;
+    private long groups;
+    private long events;
+    private long qna;
 }
